test(weave): fold failure-eviction tests into existing cache test file
Move the two new tests into tests/trace/test_wave_client_file_cache.py
and drop the WB-31070 ticket-name comment from weave_client.py.
